split source
- split source of plugin into multiple files
- move all classes under Jekyll::AsciiDoc module
- rename AsciiDocConverter to Converter
- rename AsciiDocHeaderIntegrator to Integrator
- don't use FQCN to refer to types in Jekyll::AsciiDoc module
- define empty modules using ::Module.new
- use FQCN for all types in test suite
- switch constants to pascal case
